Handle rotctld errors better

+++ src-interface/recorder/tracking_widget.cpp
@@ -177,7 +177,7 @@ namespace satdump
             if (ImGui::BeginCombo("###horizonsselectcombo", horizonsoptions[current_horizons].second.c_str()))
             {
                 ImGui::InputTextWithHint("##horizonssatellitetracking", u8"\uf422   Search", &horizonssearchstr);
-                for (int i = 0; i < horizonsoptions.size(); i++)
+                for (int i = 0; i < (int)horizonsoptions.size(); i++)
                 {
                     bool show = true;
                     if (horizonssearchstr.size() != 0)
@@ -199,7 +199,7 @@ namespace satdump
             if (ImGui::BeginCombo("###satelliteselectcombo", satoptions[current_satellite].c_str()))
             {
                 ImGui::InputTextWithHint("##searchsatellitetracking", u8"\uf422   Search", &satsearchstr);
-                for (int i = 0; i < satoptions.size(); i++)
+                for (int i = 0; i < (int)satoptions.size(); i++)
                 {
                     bool show = true;
                     if (satsearchstr.size() != 0)
